+++ mumbledj/__init__.py
@@ -1,8 +1,0 @@
-#----------------------#
-#       MumbleDJ       #
-#  By Matthieu Grieger #
-#----------------------#-------------#
-# __init__.py                        #
-# Currently empty, and will probably #
-# stay that way. :)                  #
-#------------------------------------#