+++ mumbledj/config.lua
@@ -1,0 +1,24 @@
+-------------------------
+--      MumbleDJ       --
+-- By Matthieu Grieger --
+-------------------------------------------------
+-- config.lua                                  --
+-- This is where all the configuration options --
+-- for the bot can be set.                     --
+-------------------------------------------------
+
+-------------------------
+-- GENERAL CONFIGURATION
+-------------------------
+
+-- Bot username
+-- DEFAULT VALUE: "MumbleDJ"
+BOT_USERNAME = "MumbleDJ"
+
+-- Default channel
+-- DEFAULT VALUE: "Bot Testing"
+DEFAULT_CHANNEL = "Bot Testing"
+
+-- Command prefix
+-- DEFAULT VALUE: "!"
+COMMAND_PREFIX = "!"